renmaed the src package to yamlator to ensure that when installing the package the code goes into its relevant package directory
