+++ tests/validators/test_builtin_type_validator.py
@@ -3,10 +3,10 @@ import unittest
 from .base import BaseValidatorTest
 from parameterized import parameterized
 
-from src.types import Data, Rule
-from src.types import RuleType
-from src.types import SchemaTypes
-from src.validators import BuiltInTypeValidator
+from yamlator.types import Data, Rule
+from yamlator.types import RuleType
+from yamlator.types import SchemaTypes
+from yamlator.validators import BuiltInTypeValidator
 
 
 class TestBuiltInTypeValidator(BaseValidatorTest):
