+++ tests/validators/test_enum_type_validator.py
@@ -3,12 +3,12 @@ import unittest
 from .base import BaseValidatorTest
 from parameterized import parameterized
 
-from src.types import Data
-from src.types import EnumItem
-from src.types import RuleType
-from src.types import YamlatorEnum
-from src.types import SchemaTypes
-from src.validators import EnumTypeValidator
+from yamlator.types import Data
+from yamlator.types import EnumItem
+from yamlator.types import RuleType
+from yamlator.types import YamlatorEnum
+from yamlator.types import SchemaTypes
+from yamlator.validators import EnumTypeValidator
 
 
 string_enum = {
