+++ tests/validators/test_list_validator.py
@@ -6,9 +6,9 @@ from typing import Any
 from unittest.mock import MagicMock
 from parameterized import parameterized
 
-from src.validators import ListValidator
-from src.types import RuleType
-from src.types import SchemaTypes
+from yamlator.validators import ListValidator
+from yamlator.types import RuleType
+from yamlator.types import SchemaTypes
 
 
 class TestListValidator(BaseValidatorTest):
