+++ tests/validators/test_optional_validator.py
@@ -6,8 +6,8 @@ from unittest.mock import Mock
 from unittest.mock import patch
 from parameterized import parameterized
 
-from src.types import Data
-from src.validators import OptionalValidator
+from yamlator.types import Data
+from yamlator.validators import OptionalValidator
 
 
 class TestOptionalValidator(BaseValidatorTest):
@@ -18,7 +18,7 @@ class TestOptionalValidator(BaseValidatorTest):
         ('with_optional_and_none_data', False, None, 0),
         ('with_required_and_none_data', True, None, 1),
     ])
-    @patch('src.validators.Validator.validate')
+    @patch('yamlator.validators.Validator.validate')
     def test_optional_validator(self, name: str, is_required: bool, data: Data,
                                 next_validator_call_count: int,
                                 mock_parent_validator: Mock):
