+++ tests/validators/test_regex_validator.py
@@ -3,10 +3,10 @@ import unittest
 from .base import BaseValidatorTest
 from parameterized import parameterized
 
-from src.types import Data
-from src.types import RuleType
-from src.types import SchemaTypes
-from src.validators import RegexValidator
+from yamlator.types import Data
+from yamlator.types import RuleType
+from yamlator.types import SchemaTypes
+from yamlator.validators import RegexValidator
 
 
 class TestRegexValidator(BaseValidatorTest):
