+++ tests/validators/test_validate_yaml.py
@@ -2,14 +2,14 @@ import unittest
 
 from parameterized import parameterized
 
-from src.types import Data
-from src.types import Rule
-from src.types import RuleType
-from src.types import EnumItem
-from src.types import YamlatorEnum
-from src.types import YamlatorRuleset
-from src.types import SchemaTypes
-from src.validators import validate_yaml
+from yamlator.types import Data
+from yamlator.types import Rule
+from yamlator.types import RuleType
+from yamlator.types import EnumItem
+from yamlator.types import YamlatorEnum
+from yamlator.types import YamlatorRuleset
+from yamlator.types import SchemaTypes
+from yamlator.validators import validate_yaml
 
 
 def create_flat_schema():
