+++ yamlator/__init__.py
@@ -1,0 +1,2 @@
+from yamlator.validators import validate_yaml
+from yamlator.cmd import validate_yaml_data_from_file
