+++ yamlator/validators/enum_type_validator.py
@@ -1,8 +1,8 @@
 from collections import deque
 
-from src.types import Data
-from src.types import RuleType
-from src.types import SchemaTypes
+from yamlator.types import Data
+from yamlator.types import RuleType
+from yamlator.types import SchemaTypes
 from .base_validator import Validator
