+++ yamlator/validators/map_validator.py
@@ -1,6 +1,6 @@
-from src.types import Data
-from src.types import RuleType
-from src.types import SchemaTypes
+from yamlator.types import Data
+from yamlator.types import RuleType
+from yamlator.types import SchemaTypes
 from .base_validator import Validator
