+++ yamlator/validators/regex_validator.py
@@ -1,7 +1,7 @@
-from src.types import Data
-from src.types import RuleType
-from src.types import SchemaTypes
-from src.violations import RegexTypeViolation
+from yamlator.types import Data
+from yamlator.types import RuleType
+from yamlator.types import SchemaTypes
+from yamlator.violations import RegexTypeViolation
 from .base_validator import Validator
